+++ package.json
@@ -13,7 +13,7 @@
 	"dependencies": {
 		"@fal-ai/serverless-client": "^0.6.0",
 		"@fal-ai/serverless-proxy": "^0.6.0",
-		"@tldraw/tldraw": "^2.0.0-canary.85c5210cd1a6",
+		"@tldraw/tldraw": "canary",
 		"@vercel/analytics": "^1.1.1",
 		"next": "14.0.3",
 		"react": "^18",
@@ -31,7 +31,6 @@
 		"postcss": "^8",
 		"prettier": "^3.1.0",
 		"prettier-plugin-organize-imports": "^3.2.4",
-		"tailwindcss": "^3.3.0",
 		"typescript": "^5"
 	}
 }
